requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic==2.4.2
sqlalchemy==2.0.23
sentence-transformers==2.2.2
requests==2.31.0
python-multipart==0.0.6
numpy==1.24.3
